pyproject.toml
[tool.poetry]
name = "habits-backend"
version = "0.1.0"
description = ""
authors = ["KuligKamil <kamil.marek.kulig@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
strawberry-graphql = {extras = ["debug-server"], version = "^0.171.1"}
Faker = "^18.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
